Project overhaul for using CMake.
Officially version 0.1.0!
