+++ include/utils/string_to_vec.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <sstream>
+#include <vector>
+
+#ifndef FUNCTION_STR_TO_VEC_INCLUDED
+#define FUNCTION_STR_TO_VEC_INCLUDED
+
+std::vector<std::string> string_to_vec(std::string s, std::string delimiter = " ");
+
+#endif
