+++ src/containers/dump_data_container.cpp
@@ -1,6 +1,4 @@
-#pragma once
-
-#include "../include/dump_data_container.h"
+#include "containers/dump_data_container.h"
 
 dump_data_container::dump_data_container(std::vector<double> t_vec, std::vector<int> ac_vec,
                                          std::vector<std::vector<atom>> fa_vec)
@@ -11,6 +9,3 @@ dump_data_container::dump_data_container(std::vector<double> t_vec, std::vector<
 std::vector<double> dump_data_container::get_timestep_vec() { return timestep_vec; }
 
 std::vector<int> dump_data_container::get_atomscount_vec() { return atomscount_vec; }
-
-// std::vector<boost::geometry::index::rtree<atom, boost::geometry::index::rstar<16>>>
-// dump_data_container::get_frame_atoms_vec() { return frame_atoms_vec; }
